slowly extracting graphics components
