+++ g_src/render/init.cpp
@@ -30,7 +30,7 @@ using std::string;
 
 #include "../util/g_basics.hpp"
 
-#include "../texture/texture_handler.hpp"
+#include "../graphics/texture_handler.hpp"
 
 #include "../render/graphics.hpp"
